Conflicts:
	src/Orchard/ContentManagement/DefaultContentManager.cs
	src/Orchard/ContentManagement/Handlers/ContentHandler.cs
	src/Orchard/Orchard.Framework.csproj

+++ src/Orchard.Web/Modules/Orchard.OutputCache/Filters/OutputCacheFilter.cs
@@ -19,10 +19,12 @@ using Orchard.Mvc.Filters;
 using Orchard.Services;
 using Orchard.Themes;
 using Orchard.UI.Admin;
+using Orchard.UI.Notify;
 using Orchard.Utility.Extensions;
 using System.Collections.Specialized;
 using Orchard.OutputCache.ViewModels;
 using Orchard.UI.Admin.Notification;
+using Orchard.DisplayManagement.Shapes;
 
 namespace Orchard.OutputCache.Filters {
     public class OutputCacheFilter : FilterProvider, IActionFilter, IResultFilter {
@@ -38,7 +40,6 @@ namespace Orchard.OutputCache.Filters {
         private readonly ISignals _signals;
         private readonly ShellSettings _shellSettings;
         private readonly ICacheControlStrategy _cacheControlStrategy;
-        private readonly INotificationManager _notificationManager;
 
         TextWriter _originalWriter;
         StringWriter _cachingWriter;
@@ -57,9 +58,8 @@ namespace Orchard.OutputCache.Filters {
             ICacheService cacheService,
             ISignals signals,
             ShellSettings shellSettings,
-            ICacheControlStrategy cacheControlStrategy,
-            INotificationManager notificationManager 
-            ) {
+            ICacheControlStrategy cacheControlStrategy) {
+
             _cacheManager = cacheManager;
             _cacheStorageProvider = cacheStorageProvider;
             _tagCache = tagCache;
@@ -71,7 +71,6 @@ namespace Orchard.OutputCache.Filters {
             _signals = signals;
             _shellSettings = shellSettings;
             _cacheControlStrategy = cacheControlStrategy;
-            _notificationManager = notificationManager;
 
             Logger = NullLogger.Instance;
         }
@@ -228,7 +227,8 @@ namespace Orchard.OutputCache.Filters {
             var parameters = new Dictionary<string, object>(filterContext.ActionParameters);
 
             foreach (var key in queryString.AllKeys) {
-                if (key == null) continue;
+                if (key == null || (_varyQueryStringParameters != null
+                    && !_varyQueryStringParameters.Contains(key))) continue;
 
                 // ignore pages with the RefreshKey
                 if (String.Equals(RefreshKey, key, StringComparison.OrdinalIgnoreCase)) {
@@ -351,7 +351,9 @@ namespace Orchard.OutputCache.Filters {
             }
 
             // don't cache the result if there were some notifications
-            if (_notificationManager.GetNotifications().Any()) {
+            var hasNotifications = !String.IsNullOrEmpty(Convert.ToString(filterContext.Controller.TempData["messages"]));
+            if (hasNotifications) {
+                Logger.Debug("Not caching: notifications present");
                 return;
             }
 
@@ -379,10 +381,8 @@ namespace Orchard.OutputCache.Filters {
 
             Logger.Debug("Cache item added: " + _cacheItem.CacheKey);
 
-            // remove only the current version of the page
-            _cacheService.RemoveByTag(_cacheKey);
-
-            // add data to cache
+            // update the cached data
+            _cacheStorageProvider.Remove(_cacheKey);
             _cacheStorageProvider.Set(_cacheKey, _cacheItem);
 
             // add to the tags index
@@ -424,25 +424,22 @@ namespace Orchard.OutputCache.Filters {
             Logger.Debug("Redirect on POST");
             var redirectUrl = redirectResult.Url;
 
-            if (!VirtualPathUtility.IsAbsolute(redirectUrl)) {
-                var applicationRoot = new UrlHelper(filterContext.HttpContext.Request.RequestContext).MakeAbsolute("/");
-                if (redirectUrl.StartsWith(applicationRoot, StringComparison.OrdinalIgnoreCase)) {
-                    redirectUrl = "~/" + redirectUrl.Substring(applicationRoot.Length);
-                    redirectUrl = VirtualPathUtility.ToAbsolute(redirectUrl);
-                }
+            if (filterContext.HttpContext.Request.IsLocalUrl(redirectUrl)) {
+                var helper = new UrlHelper(filterContext.HttpContext.Request.RequestContext);
+                var absolutePath = new Uri(helper.MakeAbsolute(redirectUrl)).AbsolutePath;
+
+                // querystring invariant key
+                var invariantCacheKey = ComputeCacheKey(
+                    _shellSettings.Name,
+                    absolutePath,
+                    () => _workContext.CurrentCulture,
+                    _themeManager.GetRequestTheme(filterContext.RequestContext).Id,
+                    null
+                    );
+
+                // remove all cached version of the same page
+                _cacheService.RemoveByTag(invariantCacheKey);
             }
-
-            // querystring invariant key
-            var invariantCacheKey = ComputeCacheKey(
-                _shellSettings.Name,
-                redirectUrl,
-                () => _workContext.CurrentCulture,
-                _themeManager.GetRequestTheme(filterContext.RequestContext).Id,
-                null
-                );
-
-            // remove all cached version of the same page
-            _cacheService.RemoveByTag(invariantCacheKey);
 
             // adding a refresh key so that the redirection doesn't get restored
             // from a cached version on a proxy
@@ -495,7 +492,6 @@ namespace Orchard.OutputCache.Filters {
                 response.Cache.SetMaxAge(maxAge);
             }
 
-            response.Cache.VaryByParams["*"] = true;
             response.DisableUserCache();
 
             // keeping this examples for later usage
@@ -510,7 +506,10 @@ namespace Orchard.OutputCache.Filters {
                 }
             }
 
-            if (_varyQueryStringParameters != null) {
+            if (_varyQueryStringParameters == null) {
+                response.Cache.VaryByParams["*"] = true;
+            }
+            else {
                 foreach (var queryStringParam in _varyQueryStringParameters) {
                     response.Cache.VaryByParams[queryStringParam] = true;
                 }
@@ -522,14 +521,7 @@ namespace Orchard.OutputCache.Filters {
         }
 
         private string ComputeCacheKey(ControllerContext controllerContext, IEnumerable<KeyValuePair<string, object>> parameters) {
-            var url = controllerContext.HttpContext.Request.RawUrl;
-            if (!VirtualPathUtility.IsAbsolute(url)) {
-                var applicationRoot = new UrlHelper(controllerContext.HttpContext.Request.RequestContext).MakeAbsolute("/");
-                if (url.StartsWith(applicationRoot, StringComparison.OrdinalIgnoreCase)) {
-                    url = "~/" + url.Substring(applicationRoot.Length);
-                    url = VirtualPathUtility.ToAbsolute(url);
-                }
-            }
+            var url = controllerContext.HttpContext.Request.Url.AbsolutePath;
             return ComputeCacheKey(_shellSettings.Name, url, () => _workContext.CurrentCulture, _themeManager.GetRequestTheme(controllerContext.RequestContext).Id, parameters);
         }
